+++ src/Sylius/Bundle/CartBundle/spec/Sylius/Bundle/CartBundle/Controller/CartController.php
@@ -30,4 +30,9 @@ class CartController extends ObjectBehavior
     {
         $this->shouldHaveType('Sylius\Bundle\ResourceBundle\Controller\ResourceController');
     }
+
+    function it_should_extend_base_Sylius_cart_bundle_controller()
+    {
+        $this->shouldHaveType('Sylius\Bundle\CartBundle\Controller\Controller');
+    }
 }